LayoutTests/css3/color-filters/color-filter-exposed-if-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ ColorFilterEnabled=false ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that -apple-color-filter is not exposed when the feature is disabled");

shouldBeFalse("'-apple-color-filter' in document.documentElement.style");
shouldBeFalse("'-apple-color-filter' in getComputedStyle(document.documentElement)");
shouldBeFalse("CSS.supports('-apple-color-filter: contrast(1)')");
shouldBeFalse("CSS.supports('-apple-color-filter: inherit')");

</script>
</body>
</html>
